refactor: extract Flask routes into Blueprint modules
Split web-tool.py routes into four focused Blueprint modules:
- routes/mirror_links.py: Mirror links routes and helpers
- routes/mirror_favicons.py: Favicon routes, json_response, validate_favicons
- routes/debug.py: Debug and clipboard proxy routes
- routes/javascript.py: Dynamic JS serving routes

web-tool.py reduced to 43 lines (app creation, blueprint registration,
before_request hook, logging). Added TestDeduplicateVariants and
TestNormalizeNetloc/TestGetFirstPathSegment tests.
