+++ routes/__init__.py
@@ -1,0 +1,1 @@
+"""Flask route blueprints for the web-tool application."""
